# openapi.yaml
openapi: 3.0.3
info:
  title: Liquor Distribution OS API
  version: 1.0.0
  description: |
    REST API for the Liquor Distribution Operating System.
    Covers customers, orders, invoices, collections, deliveries, inventory, and sync.
    All mutating endpoints require an `Idempotency-Key` header.
servers:
  - url: https://api.example.com/v1
    description: Production
  - url: https://staging.api.example.com/v1
    description: Staging
security:
  - bearerAuth: []
...
paths:
...
  /invoices:
    get:
      tags: [Invoices]
      parameters:
        - { in: query, name: customer_id, schema: { type: string } }
        - { in: query, name: status, schema: { type: string, enum: [open, partial, paid, disputed] } }
        - { in: query, name: overdue_min_days, schema: { type: integer } }
      responses: { '200': { description: ok } }
  /invoices/{id}:
    parameters:
      - { in: path, name: id, required: true, schema: { type: string } }
    get:
      tags: [Invoices]
      responses: { '200': { description: ok } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...